Add Point3 and Color type aliases for Vec3
Given that this type will be reused for both geometry and color, and
it's possible that we may be computing color from vector values and vice
versa, it makes sense to have a `Color` type alias here instead of a
separate struct with `Into` and `From` implementations.

The case for `Vec3` and `Point3` is less clear, but given that defining
a point requires only one vector, and defining a line requires two,
there should be minimal confusion between vectors and points in the
particular code we will be writing, so we'll add a type alias for that
as well.