Add Ray math helper type
This type models the light rays we will be emitting into the scene using
the function:

``
P(t) = A + tb
```

where _P_ is a 3D position along a line in 3D, _A_ is the ray origin,
and _b_ is the ray direction. By plugging _t_, we can perform a linear
interpolation of points along the ray.