Use Ray to generate 16:9 image of blue-to-white gradient
This commit changes `main.rs` to output a 16:9 image. We also cast one
ray to each pixel of the image, and using the image width and height
combined with a new `compute_ray_color()` linear interpolation function,
we linearly interpolate between the colors blue and white based on the
screen Y coordinate of the current pixel, creating a top-to-bottom
gradient of blue to white.

This soft gradient will be used as our scene background for the next
several chapters.

Note that our per-pixel rays are all configured to approximately hit the
pixel centers. We won't worry about exactness for now because we'll add
anti-aliasing later.