Render red sphere in center of image
This commit adds a `ray_hits_sphere()` function which is able to perform
an intersection test between a ray and a given sphere with a certain
center point and radius. We also hard-code a red sphere in the center of
the image, thereby creating our first (crude) ray-traced image.

Note that there are multiple issues with our current implementation,
most notably that we allow solutions of _t < 0_ to work fine. If you
change the sphere's center point `z = 1`, you will get the exact same
image because you can see the objects _behind_ you. We will fix those
issues later.