Render sphere color according to its normal vectors
In order to shade an object, its surface normals must be known. A
surface normal is a vector perpendicular to the surface at the point of
intersection. In our particular implementation, we assume that normal
vectors are of unit length (though we aren't statically enforcing that
in the code currently) because it's more convenient for shading.

Since we don't have any light sources implemented in this code yet, we
just visualize the normals using a color map (we transform the x, y, and
z coordinates of the current surface normal to RGB values).

In the process of working on this, we modified `ray_hits_sphere()` to
return the actual `f64` hit point of the ray, if any, rather than just
a `bool` of whether the ray hit the sphere or not.