Simplify ray-sphere intersection code
It turns out that a vector dotted with itself is equal to the squared
length of that vector, so we can replace all instances of that using our
convenient `.len_squared()` method.

Additionally, the equation for `b` has a factor of `2` in it that we can
eliminate using the quadratic equation. After switching out `b` for a
`half_b`, we can simplify the discriminant equation to this instead:

```
discriminant = half_b^2 - a * c
```