Add surface side determination to HitRecord and Sphere
We need to support two-sided surfaces that can be rendered differently
on each face, e.g. a sheet of paper, or objects that have both an inside
and an outside, like glass balls. To facilitate this, we need to be able
to determine which side of the surface our ray is intersecting with when
we color it.

There are two possible options for surface normals design-wise: they can
be modeled either as always pointing out or always pointing in. In our
particular raytracer, we choose to have normals always pointing out. We
can compare our outward normal with the incoming ray to determine
whether it's striking the front face or back side of the surface.