Replace ray_hits_sphere() with Hittable list, render ground
This commit finally replaces our old `ray_hits_sphere()` function with
the `Hittable` and `Vec<Box<dyn Hittable>>` abstraction we had created
earlier, which is composable and scalable to multi-object scenes. With
it, we are able to rapidly define two spheres: a large sphere which
serves as our ground and a smaller sphere which serves as the subject
of our image.