Add Camera helper struct
This `Camera` structure will keep track of the viewport data and handle
scene sampling tasks. To start with, we add a `ray_at()` method which
computes a ray corresponding to a given _u_ and _v_ coordinates (aka the
"screen X" and "screen Y" pixel coordinates in our output image). This
should simplify some of the `main.rs` loop.