Update compute_ray_color() to recurse with random ray direction
This commit changes `compute_ray_color()` to model a highly simplistic
form of diffuse shading by having it recurse with a new randomly
generated ray direction, thereby simulating the chaotic light ray
bounces on matte surfaces.

However, we haven't added a base case to this recursive function that
limits how many times we are allowed to recurse. In practical CS terms,
this means that this function will recurse deeper and deeper forever
until it blows the stack. And in practical optics terms, having photons
bounce around forever without ever being absorbed by the material is
physically inaccurate anyway.

Let's add that base case in the next commit in the form of a maximum
depth parameter.