Set max bounce depth of 50, produce our first diffuse image
Here we limit the maximum number of child rays such that once we exceed
the ray bounce limit, no more light is gathered. At that point, the ray
is considered to be either absorbed by the material or thrown away into
empty space.

As an aside: it's taking a longer and longer time to produce new images
in debug mode. Compiling and running in release mode helps matters
dramatically.