Add gamma correction for accurate color intensity
The previous picture we produced was unfortunately very dark, despite
our spheres being 50% reflective (meaning they absorb half the energy on
each bounce). In real life, they should appear to look light grey with
soft shadows, not dark grey with muddled dark splotches.

This isn't a bug in our algorithm; it seems to be working as intended.
Rather, the issue is that most image viewers (PPM or otherwise) assumet
that the image has been gamma corrected before serialized to bytes.

To do gamma correction, one needs to raise the R, G, and B color
components to the power `1 / gamma` before writing to disk. In our case,
for a simple approximation, we will go with `gamma = 2.0`, which results
in a power of 1/2, which is just a `sqrt()`.

The resulting image is lighter than before and looks much nicer. Look at
those shadows! Speaking of, though, the shadows are still looking a bit
splotchy. Let's correct that in the next commit.