Raise min t value to 0.001 to fix shadow acne
It turns out that the shadow splotchiness is coming from the reflected
rays not hitting the object precisely at `t = 0.0` but instead at
`t = -0.000_000_1` or `t = -0.000_000_1` or some such, due to IEEE
floating-point number limitations.

To fix this, we can simply raise the minimum _t_ value to 0.001 to
ignore those ray hits that are very close to zero. The produced shadows
look much nicer after that!