Add Vec3::random_unit() for true Lambertian distribution
The previous approach of using `Vec3::random_in_unit_sphere()` in the
process of generating random bounce rays was something of a lazy hack.
The previous reflection model had a distribution of rays that scaled by
the `cos^3(ϕ)`, where ϕ is the angle relative to the normal.

For better physical accuracy, we should be aiming for a Lambertian
distribution instead, which scales by the `cos(ϕ)` by comparison. True
Lambertian has a higher probability of ray scattering that is close to
the surface normal, unlike our previous approach, but it's much more
uniform.

We do this by picking points on the surface of the unit sphere and
normalizing them, thereby generating a random unit vector. We replace
our existing `Vec3::random_in_unit_sphere()` constructor with our new
drop-in replacement `Vec3::random_unit()`, and we now have true
Lambertian diffuse reflection!

Ultimately, the appearance of the final image is little changed, but the
shadows look noticeably softer and both spheres are a slightly lighter
shade of grey. This is due to the more uniform scattering of light
rays, thereby ensuring more rays end up bouncing toward the camera.