Add Material trait abstraction
This new `Material` trait should be useful for representing the physical
surface properties of objects with different materials.

There are generally two possible design approaches for managing
materials: we could have a single universal material with lots of
parameters, or we could have many kinds of materials which all implement
some kind of `Material` trait which unifies their behavior. We choose
the latter approach.