Remove struct derives, implement Lambertian material
In this commit, we implement a `Lambertian` material which reuses the
logic we currently have hard-coded in `src/main.rs`. For the Lambertian
case we already have, it can either (1) scatter always and attenuate by
its reflectance _R_, or (2) it can scatter with no attenuation but
absorb the fraction _1 - R_ of the rays, or (3) it could be a mixture of
those strategies. In this commit, we went with the first approach for
simplicity.

We also add a `material` field to `HitRecord`, which thereby required
some further modifications to accommodate. Note that we haven't updated
the renderer to actually make use of materials yet; that will come in a
later commit.

Unfortunately, there is no way to bound `Material` by `PartialEq` or
`Clone` and have it still be object-safe, so we remove these derives
from most of our structs.