Add fuzzy reflection to Metallic material
This commit adds an additional "fuzzy" parameter (a double between 0.0
and 1.0, where 0.0 is no perturbation) to the `Metallic` material which
randomizes the reflected direction using a small sphere and choosing a
new endpoint for the ray.

The larger this sphere, the fuzzier the reflections will be. The catch
is that for big spheres or grazing rays, we may end up scattering below
the surface. We can just have the surface absorb those.

To test this out, we set a fuzziness factor of 0.3 to the metallic
sphere on the left and 1.0 to the metallic sphere on the right.