Deal with internal reflection in dielectrics
The reason why some pixels came up with extreme and invalid color values
with the `Dielectric` material is because when the incoming ray hits the
surface and splits into both a reflected and refracted ray, once the
refracted ray is inside the material with the higher refractive index,
there is no real solution to Snell's Law, and thus there is no
refraction possible.

To handle this, we decide to first check for whether a solution to
Snell's Law exists or not for the particular incoming ray. If a solution
exists, we refract; otherwise, we reflect.

Solid dielectric objects in practice usually experience "total internal
reflection," where all the incoming light is reflected like a mirror.
This is why sometimes the water-air boundary acts like a perfect mirror
when you are submerged.

Currently, we're sticking with a dielectric material which tries to
refract where possible, reflecting only when absolutely necessary. For
now, we are also sticking with an attentuation of (1, 1, 1), meaning the
glass surface absorbs nothing.