Demonstrate hollow glass sphere
It turns out that you can mimic a hollow glass sphere by modeling two
spheres occupying the same space, one of which is inside the other and
has a negative radius. By doing so, the geometry is unaffected, but the
surface normal now points inward instead of outward. This can be used as
a bubble to make a hollow glass sphere.