Add rayon integration, add Send/Sync bounds
This yields a significant performance improvement over the previous
serial code (on my Dell XPS 13 9360, rendering time decreased from
~2m56s to ~0m45s). To achieve this, some `Send + Sync` bounds were added
to the various fundamental traits, i.e. `Hittable`, `Material`, `Sky`.

Note that this code leverages the previously introduced `std::iter::Sum`
implementation for `Vec3` to work. According to the docs for
`rayon::iter::ParallelIterator::sum`, if the `+` operator isn't truly
associative (as is the case for floating point numbers), then the
summation results are not fully deterministic. However, this shouldn't
be an issue for us because we already seed the initial pixel sampling
inputs with values derived from `rand::random()`, meaning it was never
deterministic to begin with.

The visual quality of the output doesn't seem to be affected anyway.