Replace Box<dyn Material> in Sphere with generic
Ideally, we should avoid redundant heap allocations and unnecessary
dynamic dispatch wherever possible. This means adding an `<M: Material>`
generic type parameter to `Sphere`, thereby obviating the need for a
trait object to store the material.

Normally, the addition of a generic type parameter would be a no-go for
us because it would disallow us from instantiating a heterogenous
collection of `Vec<Sphere<???>>` with different materials. However,
since the type signature of `Sphere` already gets erased when converted
into a `Box<dyn Hittable>`, this change is perfectly safe and can
improve performance.

A nice perk of this switch from trait objects to monomorphized generics
is that we can now safely `#[derive(Clone, Debug, PartialEq)]` for
`Sphere`, where this was previously impossible.