Add time information to Ray (useful for motion blur)
This will be useful in our later implementation of motion blur.

When we decided to ray trace, we decided that visual quality was worth
more than runtime performance. In the fuzzy reflection and defocus blur
(depth of field) effects implemented earlier, we simply took multiple
samples per pixel to achieve these effects, rather than attempting
anything fancy. Once you have taken a step down that road, the good news
is that almost all visual effects can be similarly brute-forced. Motion
blur is certainly one of those.

In a real camera, the shutter doesn't open and close instantaneously.
instead it opens and stays open for a time interval, and both the camera
and the objects may move during that time. It's really an average of
what the camera sees over the interval that we want.

So to begin, this commit adds a `time` field to `Ray` so that we can
keep track of when these rays are being emitted over time.