Add MovingSphere which interpolates coords based on time
This commit adds a new hittable `MovingSphere` struct which is able to
linearly interpolate its center coordinate from `center.0` at `time.0`
to `center.1` at `time.1`. Outside that time interval, the sphere
continues on at the same trajectory, so those times need not match up
with the camera shutter's open and close interval.

In the future, we may be able to consolidate `MovingSphere` and the
stationary `Sphere` together into one type, where a stationary sphere
simply has the same start and end position, but for now, we're keeping
the two separate, adhering closely to the book.