Add axis-aligned bounding box (AABB) type
This commit adds a simple `Aabb` (axis-aligned bounding box) type with
a built-in ray-slab intersection test. This will be useful for the
upcoming BVH (bounding volume hierarchy) implementation.

Just for some background on the "why" for creating a BVH: the ray-object
intersection test in `Hittable` is the main time bottleneck in a ray
tracer, and the time is linear, aka O(n), with the number of objects.
But it's a repeated search on the same model, so we ought to be able to
make it a logarithmic search in the spirit of binary search. Because
we're firing billions of rays into the same scene, we can do an analogue
of sorting the model, and then each ray intersection can be a sub-linear
search. This is where _bounding volume hierarchies_ or BVHs come in.

The key idea is to find a volume that fully encloses (bounds) all the
objects in the scene, usually a rectangular-shaped AABB, and then to
make it a sub-linear search, we need to subdivide it into a nested
hierarchy of bounding volumes. We can then perform some cheap ray-slab
intersection tests to efficiently drill down and identify the precise
object that the ray is hitting.