Construct bounding boxes for all Hittables
This commit adds a `Hittable::bounding_box()` trait method for computing
the axis-aligned bounding box for objects. We provide concrete
implementations for both `Sphere` and `MovingSphere`, as well as a
generic implementation for `Vec<Box<dyn Hittable>>` which utilizes
`aabb::surrounding_box()` to compute the minimum sized box capable of
physically containing the entire list of objects.