Implement core BVH tree with Hittable impl
This commit adds a basic BVH tree data structure, with the traversal
logic implemented in `Hittable::hit()`, but with the construction logic
not implemented yet. This structure is intended as a more efficient
container for `Hittable`s than our current `Vec`, making use of the
bounding volume hierarchy (BVH) design pattern described in detail in
the previous commit message.

The more complicated aspect of the BVH tree design, the tree-building
logic itself, will come in the next commit.