Add tree-building logic to Bvh
The basic strategy for building a BVH from a `Vec<Box<dyn Hittable>>`
goes something like this:

1. Randomly choose an axis (X, Y, or Z).
2. Sort the objects according their bounding boxes using that axis.
3. Split the list of objects in two and put half in each subtree.

One limitation of this approach is that the world cannot be empty
(cannot contain zero elements). There is no way to construct a
meaningful hierarchy from an empty list, so I've decided to add a
dependency on `anyhow` and make the `Bvh::new()` constructor fallible.

It is surprisingly easy to make this process parallel by leveraging some
nice bits and pieces from `rayon`. For one, the `ParallelSliceMut` trait
gives us a `.par_sort_unstable_by()` method which should be more
efficient than `.sort_unstable_by()` from the standard library (it seems
to have a sequential threshold built into it already, which should make
it fall back to a single-threaded sort for small input sets, for
efficiency's sake). Likewise, we can leverage `rayon::join()` to
potentially compute the subtrees in parallel.